# swagger.yaml
swagger: "2.0.0"
info:
  title: New Inland API
  description: New Inland APIs
  version: 0.1.9
...
basePath: /
schemes:
  - http
  - https
paths:
...
  /single_user_data/{id}:
    put:
      tags:
        - "User"
      summary: "Update Single User"
      description: "Update Single User"
      operationId: "findByIdAndUpdate"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the user that needs to be updated"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "Add New User"
          required: true
          schema:
            $ref: "#/definitions/Userupdate"
      responses:
        default:
          description: "Successfull Opearation"
      security:
        - api_key: []
...
  /delete_same_user_diffcollection/{phone}:
    delete:
      tags:
        - "User"
      summary: "Delete Data By User Phone Number"
      description: ""
      operationId: "deleteMany"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: "phone"
          in: "path"
          description: "Delete complete data from Mangoose"
          required: true
      responses:
        "200":
          description: "successful operation"
      security:
        - api_key: []
definitions:
...
  Userupdate:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
      profile_image:
        type: "string"

    xml:
      name: "User Add"